Source code/AlphaTrader/src/requirements.txt
pandas 
numpy
python-binance
httpimport
influxdb_client
urllib3
datetime
tqdm
joblib
